Redesign UI components. Update BRAM modal window
